backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
python-multipart==0.0.20
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9

# Development tools
black==24.10.0
flake8==7.1.1
isort==5.13.2
pytest==8.3.3
pytest-asyncio==0.23.7
mypy==1.13.0